DrawCircle expects 3 ints
https://docs.wxpython.org/wx.DC.html#wx.DC.DrawCircle

+++ src/dials/util/image_viewer/slip_viewer/ring_frame.py
@@ -284,7 +284,7 @@ class RingSettingsPanel(wx.Panel):
         dc.SetBrush(wx.Brush(colour, wx.TRANSPARENT))
         for (lon, lat, place, radius, colour, x_off, y_off, pdata) in data:
             (x, y) = self._pyslip.ConvertGeo2View((lon, lat))
-            dc.DrawCircle(x, y, radius * scale)
+            dc.DrawCircle(int(x), int(y), int(radius * scale))
 
     def DrawRing(self):
